Changes for VSCode

+++ Scene.cpp
@@ -193,7 +193,7 @@ Color Scene::CalculateLighting(CollisionResult result, Camera camera, int GISamp
 			Vec3d toCamera = (camera.location - result.location).normalized();
 			Vec3d halfVector = (toLight + toCamera).normalized();
 			float specularPower = clipf(Vec3d::Dot(result.normal, halfVector), 0, 1);
-			specularPower = clipf(std::powf(specularPower, 1/material->hardness), 0, 1) * material->specularPower;
+			specularPower = clipf(pow(specularPower, 1/material->hardness), 0, 1) * material->specularPower;
 
 			diffuseLight = diffuseLight + light->color * diffusePower * attenuation;
 			specularLight = specularLight + light->color * specularPower * attenuation;
